Testing Template vs Data Structure
Data Structure is more complicated to write and syntax is less clear, as well as being marginally slower in my direct test (may not be properly constructed tho).

So moving forward I will go more C++ way for managing data and classes with templates (ECS and UI and such).

+++ squelib_src/squelib.cpp
@@ -312,7 +312,7 @@ void SQUE_ConsolePrint(int lt, const char* log)
 
 uint32_t SQUE_RNG(uint32_t max)
 {
-    return pcg32_random_r(&random_t) / (UINT32_MAX / max);
+    return ((double)pcg32_random_r(&random_t) / (UINT32_MAX / (double)max));
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
